backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-nmap==0.7.1
httpx>=0.28.1
langchain>=0.2.0
langchain-google-genai>=1.0.6
jinja2==3.1.3
weasyprint>=62.0
python-dotenv==1.0.1
aiofiles==23.2.1
playwright>=1.44.0